# Improve performance of affinity/anti-affinity predicate by 20x in large clusters