@k82cn 2018-04-12
LGTM overall; performance is improved by avoiding checking all pods for each node.
I'd like to check more detail tomorrow, thanks very much for your patients.