@k82cn 2018-04-13
/lgtm

The fix is ok to me :)